move file reader to own file
